+++ packages/ripple-binary-codec/HISTORY.md
@@ -1,6 +1,7 @@
 # ripple-binary-codec Release History
 
 ## Unreleased
+- Fixed standard currency codes with lowercase and allowed symbols not decoding into standard codes.
 
 ## 1.4.1 (2022-06-02)
 - Added a clearer error message for trying to encode an invalid transaction. (Ex. With an incorrect TransactionType)